#6 rename: 유저 목록 Path 변경
