+++ src/components/layout/main-menu.tsx
@@ -39,7 +39,7 @@ const mainMenuData: IMenu[] = [
         id: "userList",
         name: "회원 목록",
         link: {
-          path: "/user/list",
+          path: "/user/user/list",
         },
       },
     ],
